refactor: 패키지 분리 및 Slack 로그 + 파일 로그 동시 적용
