+++ backend/src/main/java/com/woowacourse/momo/global/config/AspectConfiguration.java
@@ -1,5 +1,7 @@
 package com.woowacourse.momo.global.config;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.context.annotation.Bean;
@@ -7,9 +9,9 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
 
 import com.woowacourse.momo.support.logging.ExceptionLogging;
-import com.woowacourse.momo.support.logging.FileLogManager;
+import com.woowacourse.momo.support.logging.manager.FileLogManager;
 import com.woowacourse.momo.support.logging.Logging;
-import com.woowacourse.momo.support.logging.SlackLogManager;
+import com.woowacourse.momo.support.logging.manager.SlackLogManager;
 
 @Configuration
 @EnableAspectJAutoProxy
@@ -19,7 +21,7 @@ public class AspectConfiguration {
     private String logFilePath;
 
     @Bean
-    public FileLogManager logFileManager() {
+    public FileLogManager fileLogManager() {
         return new FileLogManager(logFilePath);
     }
 
@@ -30,7 +32,7 @@ public class AspectConfiguration {
 
     @Bean
     public Logging logging() {
-        return new Logging(logFileManager());
+        return new Logging(List.of(fileLogManager(), slackLogManager()));
     }
 
     @ConditionalOnExpression("${momo-logging.show:true}")
