+++ backend/src/main/java/com/woowacourse/momo/support/logging/Logging.java
@@ -1,6 +1,7 @@
 package com.woowacourse.momo.support.logging;
 
 import java.util.Arrays;
+import java.util.List;
 
 import org.aspectj.lang.JoinPoint;
 import org.slf4j.Logger;
@@ -8,22 +9,24 @@ import org.slf4j.LoggerFactory;
 
 import lombok.RequiredArgsConstructor;
 
+import com.woowacourse.momo.support.logging.manager.LogManager;
+
 @RequiredArgsConstructor
 public class Logging {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(Logging.class);
 
-    private final LogManager logManager;
+    private final List<LogManager> logManagers;
 
     public void printExceptionPoint(JoinPoint joinPoint) {
         String logMessage = extractExceptionInfo(joinPoint);
         LOGGER.error(ConsolePrettier.red(logMessage));
-        logManager.writeMessage(logMessage);
+        logManagers.forEach(logManager -> logManager.writeMessage(logMessage));
     }
 
     public void printStackTrace(Exception exception, JoinPoint joinPoint) {
         LOGGER.error(ConsolePrettier.red("" + TraceExtractor.getStackTrace(exception)));
-        logManager.writeException(exception);
+        logManagers.forEach(logManager -> logManager.writeException(exception);
     }
 
     private String extractExceptionInfo(JoinPoint joinPoint) {
