+++ backend/src/main/java/com/woowacourse/momo/support/logging/manager/dto/SlackMessage.java
@@ -1,4 +1,4 @@
-package com.woowacourse.momo.support.logging;
+package com.woowacourse.momo.support.logging.manager.dto;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
